+++ drel/django/api.py
@@ -1,10 +1,10 @@
 import uuid
-from typing import Dict, Callable
+from typing import Dict, Callable, Any
 
 from django.core.mail import mail_admins
 from django.http import HttpRequest, HttpResponse, RawPostDataException
 
-from drel.core import BaseFullRequestLogBuilder, log_to_es, ResponseLog, RequestLog, config
+from drel.core import BaseFullRequestLogBuilder, log_to_es, config
 from drel.core.config import request_id_storage
 from drel.utils import to_json
 
@@ -14,42 +14,42 @@ class LoggingMiddleware:
         self.get_response = get_response
 
     def __call__(self, request: HttpRequest) -> HttpResponse:
+        if config.IGNORE_LOGGING_HANDLER(request):
+            return self.get_response(request)
+
         request_id_storage.request_id = str(uuid.uuid4())
 
+        builder = DjangoFullRequestLogBuilder(user=request.user)
+
+        builder.request_to_log(request)
+
         response = self.get_response(request)
+        builder.response_to_log(response)
 
-        if config.IGNORE_LOGGING_HANDLER(request):
-            return response
-
-        log_entry = DjangoFullRequestLogBuilder(user=request.user)(request, response)
+        log_entry = builder()
         log_to_es(log_entry)
 
         return response
 
 
 class DjangoFullRequestLogBuilder(BaseFullRequestLogBuilder):
-    def request_to_log(self, request: HttpRequest) -> RequestLog:
-        return RequestLog(
-            request.get_full_path(), get_request_data(request), get_request_headers(request)
-        )
+    def _get_request_url(self, request: Any) -> str:
+        return request.get_full_path()
 
-    def response_to_log(self, response: HttpResponse) -> ResponseLog:
-        return ResponseLog(get_response_data(response), response.status_code)
+    def _get_request_data(self, request: Any) -> Dict:
+        try:
+            return request.POST.dict() or to_json(request.body)
+        except RawPostDataException:
+            return {}
 
+    def _get_request_headers(self, request: Any) -> Dict:
+        return {key: value for key, value in request.META.items() if key.startswith("HTTP")}
 
-def get_request_data(request: HttpRequest) -> Dict:
-    try:
-        return request.POST.dict() or to_json(request.body)
-    except RawPostDataException:
-        return {}
+    def _get_response_data(self, response: Any) -> Dict:
+        return to_json(response.content)
 
-
-def get_request_headers(request: HttpRequest) -> Dict:
-    return {key: value for key, value in request.META.items() if key.startswith("HTTP")}
-
-
-def get_response_data(response: HttpResponse) -> Dict:
-    return to_json(response.content)
+    def _get_response_status(self, response: Any) -> int:
+        return response.status_code
 
 
 def mail_admins_on_es_exception(index: str, doc: Dict, exception: Exception) -> None: